refact: Add base Scraper class
Add base scraper class to ensure 'DRY' principle as much possible.
