+++ scrapers/scraper_utils.py
@@ -1,4 +1,9 @@
 import os
+from datetime import datetime
+
+MODULE_PATH = os.path.abspath(os.path.dirname(__file__))
+DATA_PATH = os.path.abspath(os.path.join(MODULE_PATH, os.pardir, 'data'))
+TIMESTAMP = datetime.now().strftime('%Y%m%d')
 
 def create_directory_if_missing(file_path):
     """
