--- tests/test_data/test_performancebike_prod_specs_20180429.csv
@@ -1,3 +1,0 @@
-Handlebar,Fork,Rear Shock,Saddle,Seatpost,Chain,Bottom Bracket,Tires,Crankset,Shifters,Stem,Rack Mounts,Wheelset,Headset,Grips/Tape,Levers,Brakes,Rear Derailleur,Front Derailleur,Cassette,Pedals,Frame
-"Marin mini riser, 15mm rise, 780mm width","RockShox Recon Silver RL 27.5"" fork, 120mm travel, compression and rebound adjustment, alloy tapered steerer, 15mm thru axle","X Fusion O2 Pro R, 190x50mm, 120mm travel, Tube-B",Marin Speed Concept,"Marin, two bolt alloy",KMC X10,External seal cartridge bearing,"Schwalbe Hans Dampf, 27.5""x2.35""","Marin Forged Alloy 1x10, 32T, 76 BCD","Shimano Deore, 10-speed",Marin 3D forged alloy,,Marin Double Wall alloy,FSA Orbit,Marin Dual Density,Shimano BR-M315,"Shimano BR-M315 hydraulic disc, 180mm/160mm rotor","Shimano Deore Shadow Plus, 10-speed",N/A,"Sunrace 10-speed, 11-42T",N/A,"Series 3 6061 aluminum frame, 27.5"" wheels, 120mm travel MultiTrac suspesnion, 135mm QR"
-"Oval Concepts 310 Ergo, 6061 alloy, 31.8mm clamp, 133mm drop, 4° sweep","EMH carbon, 1 1/8"" - 1 1/4"" tapered alloy steerer",,"Oval Concepts 300, steel rail","Kestrel EMS Pro, carbon, Ritchey clamp system","KMC X11, 11-speed","Praxis M30 BSA Bottom Bracket, Cartridge Bearings","Vittoria Zaffiro Pro, 700 x 25c, folding","Oval Concepts 500, forged 6066 arms, M30 spindle, forged Praxis 50/34T rings","Shimano 105 STI, 11-speed","Oval Concepts 313, 3D-forged 6061 stem body, +/-7°",No,"Oval Concepts 327, 700c 20/24H rims","FSA integrated, 1 1/8"" top, 1 1/4"" bottom w/ 15mm top cover",Kestrel padded,Shimano 105 STI,Tektro R540 dual-pivot,"Shimano 105, 11-speed","Shimano 105, braze-on","Shimano 105, 11-28T, 11-speed",N/A,Kestrel Enhanced Modulus Hybrid (EMH) 700K & 800K carbon fiber
